Move schedule method to common module so they accesible in different routes
